fix: fix clickable logo area
